% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{ModelEvaluation}
\alias{ModelEvaluation}
\title{ModelEvaluation}
\description{
ModelEvaluation

ModelEvaluation
}
\details{
An R6 class to evaluate copula models.
Includes tools to generate tables of evaluation metrics and visualizations
to assess model performance and dependence structure.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fit_results}}{A list of fitted copula model objects.}

\item{\code{evaluation_metrics}}{A data.table containing model evaluation metrics.}

\item{\code{plots}}{A list of visualizations for evaluating copula models.}

\item{\code{data}}{The original dataset used for fitting copulas.
Initialize the ModelEvaluation class}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelEvaluation-new}{\code{ModelEvaluation$new()}}
\item \href{#method-ModelEvaluation-generate_metrics}{\code{ModelEvaluation$generate_metrics()}}
\item \href{#method-ModelEvaluation-generate_overlay_plot}{\code{ModelEvaluation$generate_overlay_plot()}}
\item \href{#method-ModelEvaluation-clone}{\code{ModelEvaluation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelEvaluation-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelEvaluation-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelEvaluation$new(fit_results, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fit_results}}{A list of fitted copula model objects.}

\item{\code{data}}{The dataset used for fitting models, in uniform margins.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of the ModelEvaluation class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelEvaluation-generate_metrics"></a>}}
\if{latex}{\out{\hypertarget{method-ModelEvaluation-generate_metrics}{}}}
\subsection{Method \code{generate_metrics()}}{
Compute and summarize evaluation metrics for fitted copulas.
Metrics include log-likelihood, AIC, BIC, and dependence measures.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelEvaluation$generate_metrics()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.table of evaluation metrics.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelEvaluation-generate_overlay_plot"></a>}}
\if{latex}{\out{\hypertarget{method-ModelEvaluation-generate_overlay_plot}{}}}
\subsection{Method \code{generate_overlay_plot()}}{
Generate overlay plot for observed vs simulated pseudo-observations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelEvaluation$generate_overlay_plot(model_name, theme = "westeros")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_name}}{The name of the model for which to generate the plot.}

\item{\code{theme}}{name of theme for \code{echarts4r}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An echarts4r plot object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelEvaluation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelEvaluation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelEvaluation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
