% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{ModelFitter}
\alias{ModelFitter}
\title{ModelFitter}
\description{
ModelFitter

ModelFitter
}
\details{
An R6 class for fitting and evaluating copula models.
Provides a library of pre-defined copula models and tools for custom models.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{A data.table containing the dataset for modeling.}

\item{\code{data_matrix}}{A matrix of post-transformed data for copula fitting}

\item{\code{copula_library}}{A pre-defined library of copula models.}

\item{\code{fit_results}}{A list to store results of copula fits.
Initialize the ModelFitter class}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelFitter-new}{\code{ModelFitter$new()}}
\item \href{#method-ModelFitter-fit_models}{\code{ModelFitter$fit_models()}}
\item \href{#method-ModelFitter-list_models}{\code{ModelFitter$list_models()}}
\item \href{#method-ModelFitter-clone}{\code{ModelFitter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelFitter-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelFitter-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelFitter$new(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A data.table containing the dataset for modeling.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of the ModelFitter class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelFitter-fit_models"></a>}}
\if{latex}{\out{\hypertarget{method-ModelFitter-fit_models}{}}}
\subsection{Method \code{fit_models()}}{
Fits multiple copula models in a loop.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelFitter$fit_models(model_names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_names}}{A character vector of model names to fit.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.table summarizing the results for each model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelFitter-list_models"></a>}}
\if{latex}{\out{\hypertarget{method-ModelFitter-list_models}{}}}
\subsection{Method \code{list_models()}}{
Lists available copula models in the library.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelFitter$list_models()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.table summarizing the copula models.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelFitter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelFitter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelFitter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
