% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_scoring.R
\name{ModelScorer}
\alias{ModelScorer}
\title{ModelScorer}
\description{
ModelScorer

ModelScorer
}
\details{
An R6 class to score copula models on new data, perform predictions,
generate simulations, and visualize the results.

This method generates large-scale simulations by splitting the total number of samples
into smaller batches. If \code{parallel} is set to TRUE, it uses the \code{future.apply} package to
process batches in parallel. On sequential processing, batches are processed one at a time.
The function ensures that all parallel sessions are closed upon completion or in the case of an error.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fit_results}}{A list of fitted copula model objects.}

\item{\code{data}}{Scoring data}

\item{\code{score_plots}}{A list of plots visualizing scored data.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelScorer-new}{\code{ModelScorer$new()}}
\item \href{#method-ModelScorer-single_instance_prediction}{\code{ModelScorer$single_instance_prediction()}}
\item \href{#method-ModelScorer-batch_prediction}{\code{ModelScorer$batch_prediction()}}
\item \href{#method-ModelScorer-large_scale_simulation}{\code{ModelScorer$large_scale_simulation()}}
\item \href{#method-ModelScorer-conditional_prediction}{\code{ModelScorer$conditional_prediction()}}
\item \href{#method-ModelScorer-hybrid_simulation}{\code{ModelScorer$hybrid_simulation()}}
\item \href{#method-ModelScorer-importance_sampling}{\code{ModelScorer$importance_sampling()}}
\item \href{#method-ModelScorer-stress_testing}{\code{ModelScorer$stress_testing()}}
\item \href{#method-ModelScorer-visualize_single_instance_prediction}{\code{ModelScorer$visualize_single_instance_prediction()}}
\item \href{#method-ModelScorer-visualize_batch_predictions}{\code{ModelScorer$visualize_batch_predictions()}}
\item \href{#method-ModelScorer-visualize_large_scale_simulation}{\code{ModelScorer$visualize_large_scale_simulation()}}
\item \href{#method-ModelScorer-visualize_conditional_predictions}{\code{ModelScorer$visualize_conditional_predictions()}}
\item \href{#method-ModelScorer-visualize_hybrid_simulation}{\code{ModelScorer$visualize_hybrid_simulation()}}
\item \href{#method-ModelScorer-visualize_importance_sampling}{\code{ModelScorer$visualize_importance_sampling()}}
\item \href{#method-ModelScorer-visualize_stress_testing}{\code{ModelScorer$visualize_stress_testing()}}
\item \href{#method-ModelScorer-clone}{\code{ModelScorer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelScorer-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelScorer-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelScorer$new(fit_results, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fit_results}}{A list of fitted copula model objects.}

\item{\code{data}}{data for scoring a single instance of copula models}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of the ModelScorer class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelScorer-single_instance_prediction"></a>}}
\if{latex}{\out{\hypertarget{method-ModelScorer-single_instance_prediction}{}}}
\subsection{Method \code{single_instance_prediction()}}{
Generate a single instance prediction using the entire model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelScorer$single_instance_prediction(model_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_name}}{Name of the model to use.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A single joint prediction as a named list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelScorer-batch_prediction"></a>}}
\if{latex}{\out{\hypertarget{method-ModelScorer-batch_prediction}{}}}
\subsection{Method \code{batch_prediction()}}{
Generate batch predictions using the entire model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelScorer$batch_prediction(model_name, n = 100)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_name}}{Name of the model to use.}

\item{\code{n}}{Number of instances to generate.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.table of batch predictions.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelScorer-large_scale_simulation"></a>}}
\if{latex}{\out{\hypertarget{method-ModelScorer-large_scale_simulation}{}}}
\subsection{Method \code{large_scale_simulation()}}{
Perform large-scale simulations using fitted copula models.
The simulation can be processed sequentially or in parallel, with user-defined batch sizes and thread counts.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelScorer$large_scale_simulation(
  model_name,
  batches = 10,
  batch_size = 1000,
  parallel = FALSE,
  threads = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_name}}{A string specifying the name of the fitted model to use.}

\item{\code{batches}}{An integer specifying the number of batches to process. Default is 10.}

\item{\code{batch_size}}{An integer specifying the number of samples per batch. Default is 1000.}

\item{\code{parallel}}{A logical value indicating whether to use parallel processing. Default is FALSE.}

\item{\code{threads}}{An integer specifying the number of threads to use for parallel processing.
If NULL, defaults to the number of available cores minus one.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.table} containing all simulated values, with a \code{BatchID} column to indicate the batch each row belongs to.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelScorer-conditional_prediction"></a>}}
\if{latex}{\out{\hypertarget{method-ModelScorer-conditional_prediction}{}}}
\subsection{Method \code{conditional_prediction()}}{
Generate conditional predictions for a subset of variables.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelScorer$conditional_prediction(model_name, known_values, n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_name}}{Name of the model to use.}

\item{\code{known_values}}{A named list of known variable values.}

\item{\code{n}}{Number of samples to get}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.table of conditional predictions.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelScorer-hybrid_simulation"></a>}}
\if{latex}{\out{\hypertarget{method-ModelScorer-hybrid_simulation}{}}}
\subsection{Method \code{hybrid_simulation()}}{
Generate hybrid simulations (conditional + unconditional).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelScorer$hybrid_simulation(model_name, known_values, n = 100)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_name}}{Name of the model to use.}

\item{\code{known_values}}{A named list of known variable values.}

\item{\code{n}}{Number of instances to simulate.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.table of hybrid simulations.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelScorer-importance_sampling"></a>}}
\if{latex}{\out{\hypertarget{method-ModelScorer-importance_sampling}{}}}
\subsection{Method \code{importance_sampling()}}{
Perform importance sampling for the copula model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelScorer$importance_sampling(model_name, n = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_name}}{Name of the model to use.}

\item{\code{n}}{Number of samples to draw.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Weighted samples.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelScorer-stress_testing"></a>}}
\if{latex}{\out{\hypertarget{method-ModelScorer-stress_testing}{}}}
\subsection{Method \code{stress_testing()}}{
Perform stress testing by simulating extreme events.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelScorer$stress_testing(model_name, n = 100)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_name}}{Name of the model to use.}

\item{\code{n}}{Number of extreme event scenarios to simulate.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.table of extreme event simulations.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelScorer-visualize_single_instance_prediction"></a>}}
\if{latex}{\out{\hypertarget{method-ModelScorer-visualize_single_instance_prediction}{}}}
\subsection{Method \code{visualize_single_instance_prediction()}}{
Visualize single instance prediction.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelScorer$visualize_single_instance_prediction(predictions)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predictions}}{A named list of single instance predictions.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An echarts4r bar chart visualization.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelScorer-visualize_batch_predictions"></a>}}
\if{latex}{\out{\hypertarget{method-ModelScorer-visualize_batch_predictions}{}}}
\subsection{Method \code{visualize_batch_predictions()}}{
Visualize batch predictions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelScorer$visualize_batch_predictions(batch_predictions)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{batch_predictions}}{A data.table of batch predictions.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An echarts4r scatter plot visualization.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelScorer-visualize_large_scale_simulation"></a>}}
\if{latex}{\out{\hypertarget{method-ModelScorer-visualize_large_scale_simulation}{}}}
\subsection{Method \code{visualize_large_scale_simulation()}}{
Visualize large-scale simulation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelScorer$visualize_large_scale_simulation(large_scale_predictions)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{large_scale_predictions}}{A data.table of large-scale predictions.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An echarts4r density plot visualization.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelScorer-visualize_conditional_predictions"></a>}}
\if{latex}{\out{\hypertarget{method-ModelScorer-visualize_conditional_predictions}{}}}
\subsection{Method \code{visualize_conditional_predictions()}}{
Visualize conditional predictions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelScorer$visualize_conditional_predictions(conditional_predictions)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conditional_predictions}}{A data.table of conditional predictions.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An echarts4r scatter plot visualization.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelScorer-visualize_hybrid_simulation"></a>}}
\if{latex}{\out{\hypertarget{method-ModelScorer-visualize_hybrid_simulation}{}}}
\subsection{Method \code{visualize_hybrid_simulation()}}{
Visualize hybrid simulations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelScorer$visualize_hybrid_simulation(hybrid_predictions)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hybrid_predictions}}{A list with conditional and unconditional components.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of echarts4r density plots.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelScorer-visualize_importance_sampling"></a>}}
\if{latex}{\out{\hypertarget{method-ModelScorer-visualize_importance_sampling}{}}}
\subsection{Method \code{visualize_importance_sampling()}}{
Visualize importance sampling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelScorer$visualize_importance_sampling(importance_samples)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{importance_samples}}{A list with samples and weights.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An echarts4r bubble plot visualization.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelScorer-visualize_stress_testing"></a>}}
\if{latex}{\out{\hypertarget{method-ModelScorer-visualize_stress_testing}{}}}
\subsection{Method \code{visualize_stress_testing()}}{
Visualize stress testing results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelScorer$visualize_stress_testing(stress_test_results)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stress_test_results}}{A data.table of stress test results.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An echarts4r density plot visualization for extreme events.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelScorer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelScorer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelScorer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
